+++ packages/frontend/.next/server/server-reference-manifest.js
@@ -1,1 +1,0 @@
-self.__RSC_SERVER_MANIFEST="{\"node\":{},\"edge\":{},\"encryptionKey\":\"3e1mNMcI2FE59irF55E4J2NKM6AjORH+JigQ5sDYXS4=\"}"